pyproject.toml
[project]
name = "qtrader"
version = "0.1.0"
description = "Quantitative Trading Environment"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "QTrader", email = "q@qtrader.app" }]
maintainers = [{ name = "QTrader", email = "q@qtrader.app" }]

dependencies = [
    "duckdb>=1.4.0",
    "pandas>=2.3.2",
    "pyarrow>=21.0.0",
    "click>=8.0.0",
    "pydantic>=2.11.9",
    "pyyaml>=6.0",
    "pytz>=2024.1",
    "structlog>=24.4.0",
]

[project.scripts]
qtrader = "qtrader.cli:main"

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.8.0",
    "pytest-watch>=4.2.0",
    "mypy>=1.14.0",
    "ruff>=0.13.0",
    "mdformat>=0.7.21",
    "mdformat-gfm>=0.3.6",
    "mdformat-tables>=1.0.0",
    "mdformat-frontmatter>=2.0.8",
    "pandas-stubs>=2.3.2.250926",
]


# Tool configurations

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
src_paths = ["src"]

[tool.mdformat]
wrap = 500
# Do not use sequential numbering - keep consistent "1." style to match markdownlint
end_of_line = "lf"

[tool.pre-commit]
# Specify custom config location
config = ".pre_commit_hooks/pre-commit-config.yaml"
